User account

+++ src/Controller/EventController.php
@@ -47,15 +47,14 @@ class EventController extends AbstractController
     /**
      * @Route("/event/new", name="event_new")
      */
-    public function new( Request $request, EntityManagerInterface $em, UserRepository $userRepository )
+    public function new( Request $request, EntityManagerInterface $em )
     {
         $event = new Event();
         $form = $this->createForm( EventType::class, $event );
 
         $form->handleRequest( $request );
         if( $form->isSubmitted() && $form->isValid() ){
-            $owner = $userRepository->find( 1 );
-            $event->setOwner( $owner );
+            $event->setOwner( $this->getUser() );
 
             $file = $event->getPictureFile();
             $filename = $this->mediaService->upload( $file );
@@ -110,6 +109,10 @@ class EventController extends AbstractController
      */
     public function update( Event $event, Request $request, EntityManagerInterface $em )
     {
+        if( $this->getUser() !== $event->getOwner() ){
+            return $this->redirectToRoute( 'main_home' );
+        }
+
         $form = $this->createForm( EventType::class, $event );
 
         $form->handleRequest( $request );
@@ -139,6 +142,10 @@ class EventController extends AbstractController
      */
     public function remove( Event $event, EntityManagerInterface $em )
     {
+        if( $this->getUser() !== $event->getOwner() ){
+            return $this->redirectToRoute( 'main_home' );
+        }
+
         $em->remove( $event );
         $em->flush();
